+++ realm/realm-library/src/objectServer/java/io/realm/SyncConfiguration.java
@@ -470,17 +470,12 @@ public class SyncConfiguration extends RealmConfiguration {
     }
 
     /**
-     * Whether this configuration is for a query-based Realm.
-     * <p>
-     * Query-based synchronization allows a synchronized Realm to be opened in such a way that
-     * only objects queried by the user are synchronized to the device.
+     * Returns whether this configuration is for a fully synchronized Realm or not.
      *
-     * @return {@code true} to open a query-based Realm {@code false} otherwise.
-     * @deprecated use {@link #isFullySynchronizedRealm()} instead.
+     * @see Builder#fullSynchronization() for more details.
      */
-    @Deprecated
-    public boolean isPartialRealm() {
-        return isPartial;
+    public boolean isFullySynchronizedRealm() {
+        return !isPartial;
     }
 
     /**
@@ -1045,14 +1040,17 @@ public class SyncConfiguration extends RealmConfiguration {
         }
 
         /**
-         * Setting this will open a query-based Realm.
+         * Define this Realm as a fully synchronized Realm.
+         * <p>
+         * Full synchronization, unlike the default query-based synchronization, will transparently
+         * synchronize the entire Realm without needing to query for the data. This option is
+         * useful if the serverside Realm is small and all the data in the Realm should be
+         * available to the user.
          *
-         * @see #isPartialRealm()
-         * @deprecated Use {@link SyncUser#createConfiguration(String)} instead.
+         * @see #isFullySynchronizedRealm() ()
          */
-        @Deprecated
-        public SyncConfiguration.Builder partialRealm() {
-            this.isPartial = true;
+        public SyncConfiguration.Builder fullSynchronization() {
+            this.isPartial = false;
             return this;
         }